+++ app/services/toggle_additional_steps.rb
@@ -8,26 +8,40 @@ class ToggleAdditionalSteps
   def call
     return unless additional_step_rule
 
-    check_to_show_additional_step!
-    check_to_hide_additional_step!
+    check_to_show_additional_step! if additional_step.hidden?
+    check_to_hide_additional_steps! unless additional_step.hidden?
   end
 
   private
-
-  def additional_step_rule
-    step.additional_step_rule
-  end
 
   def answer
     step.answer
   end
 
-  def matching_answer?
-    additional_step_rule["required_answer"].downcase == answer.response.downcase
+  def journey_steps
+    @journey_steps ||= journey.steps
+  end
+
+  def additional_step_rule
+    step.additional_step_rule
+  end
+
+  def additional_step_id
+    additional_step_rule["question_identifier"]
   end
 
   def additional_step
-    journey.steps.find_by(contentful_id: additional_step_rule["question_identifier"])
+    journey_steps.find_by(contentful_id: additional_step_id)
+  end
+
+  def matching_answer?
+    expected_answer = additional_step_rule["required_answer"].downcase
+    case answer.response.class.name
+    when "Array"
+      answer.response.map(&:downcase).include?(expected_answer)
+    else
+      expected_answer == answer.response.downcase
+    end
   end
 
   def check_to_show_additional_step!
@@ -36,9 +50,22 @@ class ToggleAdditionalSteps
     additional_step.update(hidden: false)
   end
 
-  def check_to_hide_additional_step!
+  def check_to_hide_additional_steps!
     return if matching_answer?
 
-    additional_step.update(hidden: true)
+    recursively_toggle_additional_steps!(next_step_id: additional_step_id, hidden_state: true)
+  end
+
+  def recursively_toggle_additional_steps!(next_step_id:, hidden_state:)
+    if next_step_id
+
+      next_step = journey_steps.find_by(contentful_id: next_step_id)
+      next_step.update(hidden: hidden_state)
+
+      recursively_toggle_additional_steps!(
+        next_step_id: next_step.additional_step_rule&.fetch("question_identifier", nil),
+        hidden_state: hidden_state
+      )
+    end
   end
 end