+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - questions can be hidden from view
 - specification warning if there are incomplete tasks
 - hidden questions can be shown through a new show_additional_question field for each question
+- hidden questions can be rehidden again after changing the original answer
 
 ## [release-005] - 2021-1-19